fix: Update error handling for jruby 10
With jruby 10 there is a new error type not caught by `LoadError`. This appears
to be fallout from this https://github.com/jruby/jruby/pull/8385 refactor. This
commit updates the rescue block to handle the existing `LoadError` as well as
the `Fiddle::DLError` error. This should bridge compatability between jruby 9
and 10.