fix: Safely handle Fiddle::DLError for jruby 10
This commit updates the `Google::Cloud::Env::ComputeSMBIOS#load_product_name` to
handle a new failure method introduced with Jruby10 whereby a `Fiddle::DLError`
is raised instead of a `LoadError`. In the case the `Fiddle` gem is not loadable
this safely inspects the name of the error by looking up the class name instead
of relying on the name `Fiddle::DLError` being defined.